--- tests/models/balance/sample_response.py
@@ -1,2 +1,0 @@
-def balance_response():
-    return {"balance": 1000}
