+++ tests/sampleresponse/balance.py
@@ -1,0 +1,2 @@
+def balance_response():
+    return {"balance": 1000}
